+++ src/axom/inlet/Container.cpp
@@ -436,30 +436,40 @@ namespace detail
 {
 /*!
   *****************************************************************************
-  * \brief Adds the contents of an array to the container
+  * \brief Adds the contents of an array to the table
+  * 
+  * \return The keys that were added
   *****************************************************************************
   */
 template <typename T>
-void registerCollection(Container& container,
-                        const std::unordered_map<int, T>& collection)
+std::vector<VariantKey> registerCollection(Container& container,
+                                           const std::unordered_map<int, T>& collection)
 {
+  std::vector<VariantKey> result;
   for(const auto& entry : collection)
   {
+    result.push_back(entry.first);
     container.addPrimitive(std::to_string(entry.first), "", true, entry.second);
   }
+  return result;
 }
 
 /*!
   *****************************************************************************
-  * \brief Adds the contents of a dict to the container
+  * \brief Adds the contents of a dict to the table
+  * 
+  * \return The keys that were added
   *****************************************************************************
   */
 template <typename T>
-void registerCollection(Container& container,
-                        const std::unordered_map<VariantKey, T>& collection)
+std::vector<VariantKey> registerCollection(
+  Container& container,
+  const std::unordered_map<VariantKey, T>& collection)
 {
+  std::vector<VariantKey> result;
   for(const auto& entry : collection)
   {
+    result.push_back(entry.first);
     auto string_key = indexToString(entry.first);
     SLIC_ERROR_IF(
       string_key.find('/') != std::string::npos,
@@ -469,6 +479,7 @@ void registerCollection(Container& container,
                   "[Inlet] Dictionary key cannot be the empty string");
     container.addPrimitive(string_key, "", true, entry.second);
   }
+  return result;
 }
 
 /*!
@@ -491,55 +502,57 @@ struct PrimitiveArrayHelper<Key, bool>
    * \param [inout] container The container to add the collection to
    * \param [in] reader The Reader object to read the collection from
    * \param [in] lookupPath The path within the input file to the collection
+   * 
+   * \return The keys from the collection that was added
    *****************************************************************************
    */
-  PrimitiveArrayHelper(Container& container,
-                       Reader& reader,
-                       const std::string& lookupPath)
+  static std::vector<VariantKey> add(Container& container,
+                                     Reader& reader,
+                                     const std::string& lookupPath)
   {
     std::unordered_map<Key, bool> map;
     // Failure to retrieve a map is not necessarily an error
     reader.getBoolMap(lookupPath, map);
-    registerCollection(container, map);
+    return registerCollection(container, map);
   }
 };
 
 template <typename Key>
 struct PrimitiveArrayHelper<Key, int>
 {
-  PrimitiveArrayHelper(Container& container,
-                       Reader& reader,
-                       const std::string& lookupPath)
+  static std::vector<VariantKey> add(Container& container,
+                                     Reader& reader,
+                                     const std::string& lookupPath)
   {
     std::unordered_map<Key, int> map;
     reader.getIntMap(lookupPath, map);
-    registerCollection(container, map);
+    return registerCollection(container, map);
   }
 };
 
 template <typename Key>
 struct PrimitiveArrayHelper<Key, double>
 {
-  PrimitiveArrayHelper(Container& container,
-                       Reader& reader,
-                       const std::string& lookupPath)
+  static std::vector<VariantKey> add(Container& container,
+                                     Reader& reader,
+                                     const std::string& lookupPath)
   {
     std::unordered_map<Key, double> map;
     reader.getDoubleMap(lookupPath, map);
-    registerCollection(container, map);
+    return registerCollection(container, map);
   }
 };
 
 template <typename Key>
 struct PrimitiveArrayHelper<Key, std::string>
 {
-  PrimitiveArrayHelper(Container& container,
-                       Reader& reader,
-                       const std::string& lookupPath)
+  static std::vector<VariantKey> add(Container& container,
+                                     Reader& reader,
+                                     const std::string& lookupPath)
   {
     std::unordered_map<Key, std::string> map;
     reader.getStringMap(lookupPath, map);
-    registerCollection(container, map);
+    return registerCollection(container, map);
   }
 };
 
@@ -628,17 +641,20 @@ Verifiable<Container>& Container::addPrimitiveArray(const std::string& name,
     std::string lookupPath = (pathOverride.empty()) ? fullName : pathOverride;
     lookupPath =
       removeAllInstances(lookupPath, detail::COLLECTION_GROUP_NAME + "/");
+    std::vector<VariantKey> indices;
     if(isDict)
     {
-      detail::PrimitiveArrayHelper<VariantKey, T>(container, m_reader, lookupPath);
+      indices = detail::PrimitiveArrayHelper<VariantKey, T>::add(container,
+                                                                 m_reader,
+                                                                 lookupPath);
     }
     else
     {
-      detail::PrimitiveArrayHelper<int, T>(container, m_reader, lookupPath);
+      indices =
+        detail::PrimitiveArrayHelper<int, T>::add(container, m_reader, lookupPath);
     }
     // Copy the indices to the datastore to keep track of integer vs. string indices
-    std::vector<VariantKey> indices;
-    if(m_reader.getIndices(lookupPath, indices))
+    if(!indices.empty())
     {
       container.addIndicesGroup(indices, description);
     }